+++ java/test/mloss/roc/CurvePrimitivesBuilderTest.java
@@ -45,18 +45,18 @@ public class CurvePrimitivesBuilderTest {
     };
 
     // Correct counts based on the above data (positive label = 1)
-    public static final int[] posCounts = {
+    public static final long[] posCounts = {
         0, 1, 2, 3, 3, 3, 3, 3, 3, 3, 3, 3, 3, 4, 4, 4, 5, 6
     };
-    public static final int[] negCounts = {
+    public static final long[] negCounts = {
         0, 0, 0, 0, 1, 2, 3, 4, 5, 6, 7, 8, 9, 9, 10, 11, 11, 11
     };
 
     // Correct counts based on the above data (positive label = 3)
-    public static final int[] posCountsLabel3 = {
+    public static final long[] posCountsLabel3 = {
         0, 0, 0, 0, 1, 1, 2, 2, 2, 2, 3, 3, 4, 4, 5, 6, 6, 6
     };
-    public static final int[] negCountsLabel3 = {
+    public static final long[] negCountsLabel3 = {
         0, 1, 2, 3, 3, 4, 4, 5, 6, 7, 7, 8, 8, 9, 9, 9, 10, 11
     };
